Add files via upload
The "cirrhosis" SQL project successfully demonstrates the application of SQL for managing and analysing large-scale medical datasets. Through the execution of SQL queries, the project provided valuable insights into patient demographics, medical histories, treatment trends, and outcomes. The queries allowed for in-depth exploration of the data, uncovering patterns that are crucial for decision-making in healthcare.